Endpoints for customers to view and place their orders.
Placed orders are a simple map
of {menuItemId:quantity}, actual orders are hydrated using /userinfo and
making calls to the MenuApplication resource server.

This approach uses the OAuth2RestTemplate to relay the current token to
the MenuApplication; this way the request to MenuApplication is made as
the current user. Resource server token relay is activated by a
non-blank spring.oauth2.client.clientId (even though it's not used).
Another approach would be for the OrderApplication to obtain its own
token as itself via client_credentials grant. OrderApplication would
need a client with an authorized grant type of client_credentials and
authority of menu.read, but the MenuApplication would not know who the
user that originated the request is.